Changed retrun for optimality property to float

+++ MazeMaster/Tile.m
@@ -141,7 +141,7 @@
    return e_NONE;
 }
 
-- (int)optimality
+- (float)optimality
 {
    return _cost + _heuristic;
 }
